Changes for cognition (#157)

+++ controller/transfer/util.py
@@ -40,23 +40,16 @@ def get_upload_task_message(
     return message
 
 
-def convert_to_record_dict(
+def read_file_to_df(
     file_type: str,
-    file_name: str,
+    file_path: str,
     user_id: str,
     file_import_options: str,
     project_id: str,
-) -> Tuple[List, str]:
-    if not file_type:
-        notification.create_notification(
-            NotificationType.FILE_TYPE_NOT_GIVEN,
-            user_id,
-            project_id,
-        )
+) -> pd.DataFrame:
+    if not os.path.exists(file_path):
+        raise Exception("Couldn't locate file")
 
-        if os.path.exists(file_name):
-            os.remove(file_name)
-        raise Exception("Upload conversion error", "Upload ran into errors")
     file_type = file_type.lower()
     if file_type in ["xls", "xlsm", "xlsb", "odf", "ods", "odt"]:
         file_type = "xlsx"
@@ -68,13 +61,11 @@ def convert_to_record_dict(
     )
     try:
         if file_type in ["csv", "txt", "text"]:
-            df = pd.read_csv(file_name, **file_import_options)
+            df = pd.read_csv(file_path, **file_import_options)
         elif file_type == "xlsx":
-            df = pd.read_excel(file_name, **file_import_options)
-        elif file_type == "html":
-            df = pd.read_html(file_name, **file_import_options)
+            df = pd.read_excel(file_path, **file_import_options)
         elif file_type == "json":
-            df = pd.read_json(file_name, **file_import_options)
+            df = pd.read_json(file_path, **file_import_options)
         else:
             notification.create_notification(
                 NotificationType.INVALID_FILE_TYPE,
@@ -94,11 +85,41 @@ def convert_to_record_dict(
             project_id,
             str(e),
         )
+        raise Exception("Upload conversion error", "Upload ran into errors")
+    return df
+
+
+def convert_to_record_dict(
+    file_type: str,
+    file_name: str,
+    user_id: str,
+    file_import_options: str,
+    project_id: str,
+    column_mapping: Optional[Dict[str, str]] = None,
+) -> Tuple[List, str]:
+    if not file_type:
+        notification.create_notification(
+            NotificationType.FILE_TYPE_NOT_GIVEN,
+            user_id,
+            project_id,
+        )
+
         if os.path.exists(file_name):
             os.remove(file_name)
         raise Exception("Upload conversion error", "Upload ran into errors")
+    try:
+        df = read_file_to_df(
+            file_type, file_name, user_id, file_import_options, project_id
+        )
+    except Exception as e:
+        if os.path.exists(file_name):
+            os.remove(file_name)
+        raise e
     if os.path.exists(file_name):
         os.remove(file_name)
+
+    if column_mapping:
+        df.rename(columns=column_mapping, inplace=True)
     run_limit_checks(df, project_id, user_id)
     run_checks(df, project_id, user_id)
     check_and_convert_category_for_unknown(df, project_id, user_id)
